Re-ran the example with the newest version of the code.
Fixed a small bug in count_cells_all_scans.

Changed the layout of the GUI for segmentation to scale better.

Made it so that right clicking (without hitting the "Delete" button) can delete annotations in the segmentation GUI. Also made it so that you don't have to right click within the object, just within a small radius of the object (very helpful for deleting tiny accidental segmentations).