Slightly modified the algorithm for counting the number of cells in a scan.
The algorithm works by blurring the scan and finding local maxima. Many local maxima are detected where there are no cells, so the code makes a histogram of the image intensity at each local max. There are many many more false positive local max than true positive local max and the false positives always have lower intensity. This means the histogram is bimodal and the local max to keep are those on the right side.

The old version eliminated these false positives by identifying the location on the histogram where the slope first became positive (which is presumably after the first peak in the histogram. Generally this identified a good intensity for thresholding the local max.

This current version of the code looks for the first location where the slope of the histogram is positive, but only AFTER the maxima of the histogram. This version of the code makes the algorithm for cell detection robust to scans with high background. The old version of the code was identifying the wrong threshold for such scans.